Refactor css/style tests
